Title: 바이러스 검사, Time: 364ms, Memory: 4MB, Status: Passed - Codetree

+++ 250405/바이러스 검사/virus-detector.cpp
@@ -14,7 +14,7 @@ int main() {
     int l, s;
     cin>>l>>s;
 
-    int res = 0;
+    long long int res = 0;
 
     for(auto pp:p){
         res+=1;
